Addresses https://github.com/TanStack/router/discussions/2011 to improve performance of `Link` and `router.buildLocation`.

`matchRoutesInternal` traverses all routes in `flatRoutes`, which degrades the performance of Link with the number of routes registered. This PR adds an optimization to look up the route in `routesByPath` in constant time for absolute paths in `to`.

Rendering times of 10000 Links as described in the original discussion thread https://github.com/TanStack/router/discussions/2011#discussioncomment-10668917
| | 2 routes | 400+ routes | 400+ routes optimized
|-|-|-|-|
| hardcoded href | 150ms | 160ms | 150ms |
| interpolatePath | 180ms | 190ms | 180ms |
| buildLocation | 360ms | 1760ms ⚠️  | 320ms |
| Link component | 660ms | 2160ms ⚠️  | 590ms |