@schiller-manuel 2024-09-18
how did you measure the time?
would make sense to add this benchmark to the repo