@honzahruby 2024-09-19
> how did you measure the time? would make sense to add this benchmark to the repo

I used [this code](https://gist.github.com/honzahruby/8e189b6f9743ba78962577a8bf346c6a) in the _large-file-based_ example and measured the time to render the Home component in react dev tools. It renders 10000 links with generated IDs used as a route param.

Is there any performance benchmark setup in this repo? How about https://www.npmjs.com/package/reassure?