@schiller-manuel 2024-09-20
a performance setup as you proposed would be very good. can you set something up in a separate PR so we can compare different solutions to this perf issue?

we should also look into whether we can cache those expensive things