@honzahruby 2024-09-20
Sure I can try.

As for this problem, caching won’t help when first generating the links, you have to traverse all the routes for each link. I was playing with caching various steps for a good week before I came up with this PR 🙂