@honzahruby 2024-09-23
@schiller-manuel I set up a basic performance testing in #2398. Interestingly, the use case without `to` turned out to be the slowest one at the moment (11x slower than absolute path). With fixes applied, it's on par with the normal one. I'll update this PR when we move forward with the testing stuff.