@schiller-manuel 2024-10-12
I applied the idea in #2516
thanks for this!
if you have more perf ideas (or anything else), PRs are always welcome!